Add liquidity

+++ contracts/interfaces/IBorrow.sol
@@ -13,118 +13,106 @@ interface IBorrow {
         address from;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
-        uint128 collateralIn;
+        uint112 assetOut;
+        uint112 collateralIn;
         uint112 interestIncrease;
         uint112 cdpIncrease;
         bool isMaxCollateral;
-        uint256 maxCollateralOrDebt;
+        uint112 maxCollateralOrDebt;
         uint256 deadline;
     }
 
     struct _BorrowGivenCollateral {
-        
         IERC20 asset;
         IERC20 collateral;
         uint256 maturity;
-        
         address from;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
-        uint128 collateralLocked;
-        uint256 maxDebt;
+        uint112 assetOut;
+        uint112 collateralLocked;
+        uint112 maxDebt;
         uint256 deadline;
     }
 
-
     struct BorrowGivenCollateral {
-        
         IERC20 asset;
         IERC20 collateral;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
-        uint128 collateralLocked;
-        uint256 maxDebt;
+        uint112 assetOut;
+        uint112 collateralLocked;
+        uint112 maxDebt;
         uint256 deadline;
     }
 
     struct BorrowGivenCollateralETHAsset {
-        
         IERC20 collateral;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
-        uint128 collateralLocked;
-        uint256 maxDebt;
+        uint112 assetOut;
+        uint112 collateralLocked;
+        uint112 maxDebt;
         uint256 deadline;
     }
 
     struct BorrowGivenCollateralETHCollateral {
-        
         IERC20 asset;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
-        uint128 collateralLocked;
-        uint256 maxDebt;
+        uint112 assetOut;
+        uint112 collateralLocked;
+        uint112 maxDebt;
         uint256 deadline;
     }
 
     struct _BorrowGivenDebt {
-        
         IERC20 asset;
         IERC20 collateral;
         uint256 maturity;
-        
         address from;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
+        uint112 assetOut;
         uint112 debt;
-        uint256 maxCollateral;
+        uint112 maxCollateral;
         uint256 deadline;
     }
 
     struct BorrowGivenDebt {
-        
         IERC20 asset;
         IERC20 collateral;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
+        uint112 assetOut;
         uint112 debt;
-        uint256 maxCollateral;
+        uint112 maxCollateral;
         uint256 deadline;
     }
 
     struct BorrowGivenDebtETHAsset {
-        
         IERC20 collateral;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
+        uint112 assetOut;
         uint112 debt;
-        uint256 maxCollateral;
+        uint112 maxCollateral;
         uint256 deadline;
     }
 
     struct BorrowGivenDebtETHCollateral {
-        
         IERC20 asset;
         uint256 maturity;
         address assetTo;
         address dueTo;
-        uint128 assetOut;
+        uint112 assetOut;
         uint112 debt;
-        uint256 maxCollateral;
+        uint112 maxCollateral;
         uint256 deadline;
     }
-
 }
